feat - add support for JavaScript modules and improvements
